requirements.txt
ollama
google-generativeai

